requirements.txt
Django==2.2.10
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.0
